requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pyttsx3>=2.90
SpeechRecognition>=3.10.0
pyaudio>=0.2.13
yt-dlp>=2024.0.0
qrcode>=7.4.2
Pillow>=10.0.0

# QR scanning (choose one):
# pyzbar>=0.1.9  # Requires ZBar DLL on Windows
opencv-python>=4.8.0  # Easier on Windows
